Ajustes

+++ resources/views/personas/miembros.blade.php
@@ -3,7 +3,7 @@
         <tr>
         <th>Documento</th>
         <th>Nombre</th>
-        <th>Apellido</th>
+        <th>Parentezco</th>
         <th>Sexo</th>
         {{-- <th>Fecha Nac</th>--}}
         <th>Email</th>
@@ -20,13 +20,14 @@
     <tbody>
     @foreach($personas as $persona)
         <tr>
-            <td>{!! $persona->ci !!}</td>
-            <td>{!! $persona->nombre !!}</td>
-            <td>{!! $persona->apellido !!}</td>
-            <td>{!! $persona->sexo !!}</td>
+            <td>{!! $persona->persona_id?$persona->persona->ci:"" !!}</td>
+            <td>{!! $persona->persona_id?$persona->persona->nombre:"" !!} {!! $persona->persona_id?$persona->persona->apellido:"" !!}</td>
+            <td>{!! $persona->parentesco_id?$persona->parentesco->name:"" !!}</td>
+            <td>{!! $persona->persona_id?$persona->persona->sexo:"" !!}</td>
+
           {{--  <td>{!! $persona->fecha_nac !!}</td>--}}
-            <td>{!! $persona->email !!}</td>
-            <td>{!! $persona->celular !!}</td>
+            <td>{!! $persona->persona_id?$persona->persona->email:"" !!}</td>
+            <td>{!! $persona->persona_id?$persona->persona->celular:"" !!}</td>
             {{--<td>{!! $persona->domicilio_actual !!}</td>
             <td>{!! $persona->departamento !!}</td>
             <td>{!! $persona->ciudad !!}</td>
@@ -34,6 +35,7 @@
             <td>{!! $persona->ingreso !!}</td>
             <td>{!! $persona->discapacidad !!}</td>--}}
             <td>
+            @if(!$persona->parentesco_id == 1)
                 {!! Form::open(['route' => ['personas.destroy', $persona->id], 'method' => 'delete']) !!}
                 <div class='btn-group'>
                     <a href="{!! route('personas.show', [$persona->id]) !!}" class='btn btn-default btn-xs'><i class="glyphicon glyphicon-eye-open"></i></a>
@@ -41,6 +43,7 @@
                     {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
                 </div>
                 {!! Form::close() !!}
+            @endif
             </td>
         </tr>
     @endforeach
